Add support for using an _external_ database
i.e. a database not supplied by this chart.

Rationale: Multiple reasons:

- The bitnami charts are going away, so we cannot rely on using them
in the future
- If people use e.g. mariadb-operator (
https://github.com/mariadb-operator/mariadb-operator ), it would be
nice to be able to support that
- When deploying in a cloud environment, users may want to use a
MySQL/MariaDB instance provided by the cloud provider

The chart will still default to use the bitnami mariadb chart for
backward compatibility, so for existing users: nothing will change.